ci: add ci, refactor: remove imessage with method and add to smarts
