+++ src/WHyLL/ByteMessage/AsByteMessage.cs
@@ -7,9 +7,6 @@ public sealed class AsByteMessage(byte[] head, IEnumerable<IPair<string,byte[]>>
 {
     public IByteMessage With(byte[] newHead) => new AsByteMessage(head, attributes, body);
 
-    public IByteMessage With(params IPair<string, byte[]>[] newAttributes) =>
-        new AsByteMessage(head, Joined._(attributes, newAttributes), body);
-
     public IByteMessage With(IEnumerable<IPair<string, byte[]>> newAttributes) =>
         new AsByteMessage(head, Joined._(attributes, newAttributes), body);
