+++ src/WHyLL/Message/Lambda.cs
@@ -16,9 +16,6 @@ namespace WHyLL.Message
             message.Value.With(firstLine);
 
         public IMessage With(IEnumerable<IPair<string, string>> newParts) =>
-            this.With(newParts.ToArray());
-
-        public IMessage With(params IPair<string, string>[] newParts) =>
             this.message.Value.With(newParts);
 
         public IMessage WithBody(Stream newBody) =>
