+++ src/WHyLL/Message/MessageSmarts.cs
@@ -1,4 +1,5 @@
 using Tonga;
+using Tonga.Enumerable;
 
 namespace WHyLL.Message;
 
@@ -9,4 +10,10 @@ public static class MessageSmarts
     /// </summary>
     public static IMessage With(this IMessage msg, IText firstLine) =>
         msg.With(firstLine.AsString());
+
+    /// <summary>
+    /// Message with new headers.
+    /// </summary>
+    public static IMessage With(this IMessage msg, params IPair<string, string>[] newParts) =>
+        msg.With(new AsEnumerable<IPair<string,string>>(newParts));
 }
